refactor: split mtgo_bridge_service/client.py into discovery/commands/watch
Decompose the 452-LOC client.py into three function-based submodules
matching the gamelog_service idiom:

- discovery.py: bridge discovery / process resolution
- commands.py: one-shot command transport (futures, run/submit, fetch_*)
- watch.py: streaming challenge-watch transport (BridgeWatcher, start_watch)

client.py is now a thin re-export facade so every existing import path
(services.mtgo_bridge_service.client.*) continues to resolve, including the
private helpers (_require_bridge_path, _resolve_bridge_path, ...) used by the
package __init__ and tests. Behavior is unchanged.
